Verify integrity of keys and values using MD5-based checksums

+++ src/Fugu.Core/Format/CommitFooterRecord.cs
@@ -5,6 +5,6 @@ namespace Fugu.Format
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct CommitFooterRecord
     {
-        public ulong CommitChecksum { get; set; }
+        public uint CommitChecksum { get; set; }
     }
 }
